Rename/move a bunch of things for better package organization
