+++ squidb-processor/src/com/yahoo/squidb/processor/plugins/defaults/properties/BaseFieldPlugin.java
@@ -3,21 +3,21 @@
  * Copyrights licensed under the Apache 2.0 License.
  * See the accompanying LICENSE file for terms.
  */
-package com.yahoo.squidb.processor.plugins.properties.factory;
+package com.yahoo.squidb.processor.plugins.defaults.properties;
 
 import com.yahoo.aptutils.model.DeclaredTypeName;
 import com.yahoo.aptutils.utils.AptUtils;
 import com.yahoo.squidb.processor.data.ModelSpec;
 import com.yahoo.squidb.processor.plugins.Plugin;
-import com.yahoo.squidb.processor.plugins.properties.generators.PropertyGenerator;
+import com.yahoo.squidb.processor.plugins.defaults.properties.generators.PropertyGenerator;
 
 import java.lang.annotation.Annotation;
 
 import javax.lang.model.element.VariableElement;
 
-public abstract class PropertyGeneratorPlugin<T extends Annotation> extends Plugin {
+public abstract class BaseFieldPlugin<T extends Annotation> extends Plugin {
 
-    public PropertyGeneratorPlugin(ModelSpec<?> modelSpec, AptUtils utils) {
+    public BaseFieldPlugin(ModelSpec<?> modelSpec, AptUtils utils) {
         super(modelSpec, utils);
     }
