+++ squidb-processor/src/com/yahoo/squidb/processor/plugins/defaults/properties/FieldReferencePlugin.java
@@ -3,22 +3,22 @@
  * Copyrights licensed under the Apache 2.0 License.
  * See the accompanying LICENSE file for terms.
  */
-package com.yahoo.squidb.processor.plugins.properties.factory;
+package com.yahoo.squidb.processor.plugins.defaults.properties;
 
 import com.yahoo.aptutils.model.DeclaredTypeName;
 import com.yahoo.aptutils.utils.AptUtils;
 import com.yahoo.squidb.processor.TypeConstants;
 import com.yahoo.squidb.processor.data.ModelSpec;
-import com.yahoo.squidb.processor.plugins.properties.generators.PropertyGenerator;
-import com.yahoo.squidb.processor.plugins.properties.generators.ViewPropertyGenerator;
+import com.yahoo.squidb.processor.plugins.defaults.properties.generators.PropertyGenerator;
+import com.yahoo.squidb.processor.plugins.defaults.properties.generators.ViewPropertyGenerator;
 
 import java.lang.annotation.Annotation;
 
 import javax.lang.model.element.VariableElement;
 
-abstract class FieldReferencePropertyGeneratorFactory<T extends Annotation> extends PropertyGeneratorPlugin<T> {
+abstract class FieldReferencePlugin<T extends Annotation> extends BaseFieldPlugin<T> {
 
-    public FieldReferencePropertyGeneratorFactory(ModelSpec<?> modelSpec, AptUtils utils) {
+    public FieldReferencePlugin(ModelSpec<?> modelSpec, AptUtils utils) {
         super(modelSpec, utils);
     }
