+++ squidb-processor/src/com/yahoo/squidb/processor/plugins/defaults/properties/generators/BasicBlobPropertyGenerator.java
@@ -3,41 +3,39 @@
  * Copyrights licensed under the Apache 2.0 License.
  * See the accompanying LICENSE file for terms.
  */
-package com.yahoo.squidb.processor.plugins.properties.generators;
+package com.yahoo.squidb.processor.plugins.defaults.properties.generators;
 
-import com.yahoo.aptutils.model.CoreTypes;
 import com.yahoo.aptutils.model.DeclaredTypeName;
 import com.yahoo.aptutils.utils.AptUtils;
 import com.yahoo.squidb.processor.TypeConstants;
 import com.yahoo.squidb.processor.data.ModelSpec;
 
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import javax.lang.model.element.VariableElement;
 
 /**
- * An implementation of {@link PropertyGenerator} for handling double fields
+ * An implementation of {@link PropertyGenerator} for handling blob (byte[]) fields
  */
-public class BasicDoublePropertyGenerator extends BasicPropertyGenerator {
+public class BasicBlobPropertyGenerator extends BasicPropertyGenerator {
 
     public static List<DeclaredTypeName> handledColumnTypes() {
-        return Arrays.asList(CoreTypes.JAVA_FLOAT, CoreTypes.PRIMITIVE_FLOAT,
-                CoreTypes.JAVA_DOUBLE, CoreTypes.PRIMITIVE_DOUBLE);
+        return Collections.singletonList(TypeConstants.BYTE_ARRAY);
     }
 
-    public BasicDoublePropertyGenerator(ModelSpec<?> modelSpec, VariableElement field, AptUtils utils) {
+    public BasicBlobPropertyGenerator(ModelSpec<?> modelSpec, VariableElement field, AptUtils utils) {
         super(modelSpec, field, utils);
     }
 
     @Override
     protected DeclaredTypeName getTypeForGetAndSet() {
-        return CoreTypes.JAVA_DOUBLE;
+        return TypeConstants.BYTE_ARRAY;
     }
 
     @Override
     public DeclaredTypeName getPropertyType() {
-        return TypeConstants.DOUBLE_PROPERTY;
+        return TypeConstants.BLOB_PROPERTY;
     }
 
 }
