+++ squidb-processor/src/com/yahoo/squidb/processor/writers/ViewModelFileWriter.java
@@ -16,7 +16,7 @@ import com.yahoo.squidb.annotations.Alias;
 import com.yahoo.squidb.processor.TypeConstants;
 import com.yahoo.squidb.processor.data.ViewModelSpecWrapper;
 import com.yahoo.squidb.processor.plugins.PluginManager;
-import com.yahoo.squidb.processor.plugins.properties.generators.PropertyGenerator;
+import com.yahoo.squidb.processor.plugins.defaults.properties.generators.PropertyGenerator;
 
 import java.io.IOException;
